Floats to Doubles

+++ Detour/Include/DetourCommon.h
@@ -74,7 +74,7 @@ template<class T> inline T dtClamp(T v, T mn, T mx) { return v < mn ? mn : (v > 
 /// Returns the square root of the value.
 ///  @param[in]		x	The value.
 ///  @return The square root of the vlaue.
-float dtSqrt(float x);
+double dtSqrt(double x);
 
 /// @}
 /// @name Vector helper functions.
@@ -84,7 +84,7 @@ float dtSqrt(float x);
 ///  @param[out]	dest	The cross product. [(x, y, z)]
 ///  @param[in]		v1		A Vector [(x, y, z)]
 ///  @param[in]		v2		A vector [(x, y, z)]
-inline void dtVcross(float* dest, const float* v1, const float* v2)
+inline void dtVcross(double* dest, const double* v1, const double* v2)
 {
 	dest[0] = v1[1]*v2[2] - v1[2]*v2[1];
 	dest[1] = v1[2]*v2[0] - v1[0]*v2[2];
@@ -95,7 +95,7 @@ inline void dtVcross(float* dest, const float* v1, const float* v2)
 ///  @param[in]		v1	A Vector [(x, y, z)]
 ///  @param[in]		v2	A vector [(x, y, z)]
 /// @return The dot product.
-inline float dtVdot(const float* v1, const float* v2)
+inline double dtVdot(const double* v1, const double* v2)
 {
 	return v1[0]*v2[0] + v1[1]*v2[1] + v1[2]*v2[2];
 }
@@ -105,7 +105,7 @@ inline float dtVdot(const float* v1, const float* v2)
 ///  @param[in]		v1		The base vector. [(x, y, z)]
 ///  @param[in]		v2		The vector to scale and add to @p v1. [(x, y, z)]
 ///  @param[in]		s		The amount to scale @p v2 by before adding to @p v1.
-inline void dtVmad(float* dest, const float* v1, const float* v2, const float s)
+inline void dtVmad(double* dest, const double* v1, const double* v2, const double s)
 {
 	dest[0] = v1[0]+v2[0]*s;
 	dest[1] = v1[1]+v2[1]*s;
@@ -117,7 +117,7 @@ inline void dtVmad(float* dest, const float* v1, const float* v2, const float s)
 ///  @param[in]		v1		The starting vector.
 ///  @param[in]		v2		The destination vector.
 ///	 @param[in]		t		The interpolation factor. [Limits: 0 <= value <= 1.0]
-inline void dtVlerp(float* dest, const float* v1, const float* v2, const float t)
+inline void dtVlerp(double* dest, const double* v1, const double* v2, const double t)
 {
 	dest[0] = v1[0]+(v2[0]-v1[0])*t;
 	dest[1] = v1[1]+(v2[1]-v1[1])*t;
@@ -128,7 +128,7 @@ inline void dtVlerp(float* dest, const float* v1, const float* v2, const float t
 ///  @param[out]	dest	The result vector. [(x, y, z)]
 ///  @param[in]		v1		The base vector. [(x, y, z)]
 ///  @param[in]		v2		The vector to add to @p v1. [(x, y, z)]
-inline void dtVadd(float* dest, const float* v1, const float* v2)
+inline void dtVadd(double* dest, const double* v1, const double* v2)
 {
 	dest[0] = v1[0]+v2[0];
 	dest[1] = v1[1]+v2[1];
@@ -139,7 +139,7 @@ inline void dtVadd(float* dest, const float* v1, const float* v2)
 ///  @param[out]	dest	The result vector. [(x, y, z)]
 ///  @param[in]		v1		The base vector. [(x, y, z)]
 ///  @param[in]		v2		The vector to subtract from @p v1. [(x, y, z)]
-inline void dtVsub(float* dest, const float* v1, const float* v2)
+inline void dtVsub(double* dest, const double* v1, const double* v2)
 {
 	dest[0] = v1[0]-v2[0];
 	dest[1] = v1[1]-v2[1];
@@ -150,7 +150,7 @@ inline void dtVsub(float* dest, const float* v1, const float* v2)
 ///  @param[out]	dest	The result vector. [(x, y, z)]
 ///  @param[in]		v		The vector to scale. [(x, y, z)]
 ///  @param[in]		t		The scaling factor.
-inline void dtVscale(float* dest, const float* v, const float t)
+inline void dtVscale(double* dest, const double* v, const double t)
 {
 	dest[0] = v[0]*t;
 	dest[1] = v[1]*t;
@@ -160,7 +160,7 @@ inline void dtVscale(float* dest, const float* v, const float t)
 /// Selects the minimum value of each element from the specified vectors.
 ///  @param[in,out]	mn	A vector.  (Will be updated with the result.) [(x, y, z)]
 ///  @param[in]	v	A vector. [(x, y, z)]
-inline void dtVmin(float* mn, const float* v)
+inline void dtVmin(double* mn, const double* v)
 {
 	mn[0] = dtMin(mn[0], v[0]);
 	mn[1] = dtMin(mn[1], v[1]);
@@ -170,7 +170,7 @@ inline void dtVmin(float* mn, const float* v)
 /// Selects the maximum value of each element from the specified vectors.
 ///  @param[in,out]	mx	A vector.  (Will be updated with the result.) [(x, y, z)]
 ///  @param[in]		v	A vector. [(x, y, z)]
-inline void dtVmax(float* mx, const float* v)
+inline void dtVmax(double* mx, const double* v)
 {
 	mx[0] = dtMax(mx[0], v[0]);
 	mx[1] = dtMax(mx[1], v[1]);
@@ -182,7 +182,7 @@ inline void dtVmax(float* mx, const float* v)
 ///  @param[in]		x		The x-value of the vector.
 ///  @param[in]		y		The y-value of the vector.
 ///  @param[in]		z		The z-value of the vector.
-inline void dtVset(float* dest, const float x, const float y, const float z)
+inline void dtVset(double* dest, const double x, const double y, const double z)
 {
 	dest[0] = x; dest[1] = y; dest[2] = z;
 }
@@ -190,7 +190,7 @@ inline void dtVset(float* dest, const float x, const float y, const float z)
 /// Performs a vector copy.
 ///  @param[out]	dest	The result. [(x, y, z)]
 ///  @param[in]		a		The vector to copy. [(x, y, z)]
-inline void dtVcopy(float* dest, const float* a)
+inline void dtVcopy(double* dest, const double* a)
 {
 	dest[0] = a[0];
 	dest[1] = a[1];
@@ -200,7 +200,7 @@ inline void dtVcopy(float* dest, const float* a)
 /// Derives the scalar length of the vector.
 ///  @param[in]		v The vector. [(x, y, z)]
 /// @return The scalar length of the vector.
-inline float dtVlen(const float* v)
+inline double dtVlen(const double* v)
 {
 	return dtSqrt(v[0]*v[0] + v[1]*v[1] + v[2]*v[2]);
 }
@@ -208,7 +208,7 @@ inline float dtVlen(const float* v)
 /// Derives the square of the scalar length of the vector. (len * len)
 ///  @param[in]		v The vector. [(x, y, z)]
 /// @return The square of the scalar length of the vector.
-inline float dtVlenSqr(const float* v)
+inline double dtVlenSqr(const double* v)
 {
 	return v[0]*v[0] + v[1]*v[1] + v[2]*v[2];
 }
@@ -217,11 +217,11 @@ inline float dtVlenSqr(const float* v)
 ///  @param[in]		v1	A point. [(x, y, z)]
 ///  @param[in]		v2	A point. [(x, y, z)]
 /// @return The distance between the two points.
-inline float dtVdist(const float* v1, const float* v2)
+inline double dtVdist(const double* v1, const double* v2)
 {
-	const float dx = v2[0] - v1[0];
-	const float dy = v2[1] - v1[1];
-	const float dz = v2[2] - v1[2];
+	const double dx = v2[0] - v1[0];
+	const double dy = v2[1] - v1[1];
+	const double dz = v2[2] - v1[2];
 	return dtSqrt(dx*dx + dy*dy + dz*dz);
 }
 
@@ -229,11 +229,11 @@ inline float dtVdist(const float* v1, const float* v2)
 ///  @param[in]		v1	A point. [(x, y, z)]
 ///  @param[in]		v2	A point. [(x, y, z)]
 /// @return The square of the distance between the two points.
-inline float dtVdistSqr(const float* v1, const float* v2)
+inline double dtVdistSqr(const double* v1, const double* v2)
 {
-	const float dx = v2[0] - v1[0];
-	const float dy = v2[1] - v1[1];
-	const float dz = v2[2] - v1[2];
+	const double dx = v2[0] - v1[0];
+	const double dy = v2[1] - v1[1];
+	const double dz = v2[2] - v1[2];
 	return dx*dx + dy*dy + dz*dz;
 }
 
@@ -243,10 +243,10 @@ inline float dtVdistSqr(const float* v1, const float* v2)
 /// @return The distance between the point on the xz-plane.
 ///
 /// The vectors are projected onto the xz-plane, so the y-values are ignored.
-inline float dtVdist2D(const float* v1, const float* v2)
+inline double dtVdist2D(const double* v1, const double* v2)
 {
-	const float dx = v2[0] - v1[0];
-	const float dz = v2[2] - v1[2];
+	const double dx = v2[0] - v1[0];
+	const double dz = v2[2] - v1[2];
 	return dtSqrt(dx*dx + dz*dz);
 }
 
@@ -254,18 +254,18 @@ inline float dtVdist2D(const float* v1, const float* v2)
 ///  @param[in]		v1	A point. [(x, y, z)]
 ///  @param[in]		v2	A point. [(x, y, z)]
 /// @return The square of the distance between the point on the xz-plane.
-inline float dtVdist2DSqr(const float* v1, const float* v2)
+inline double dtVdist2DSqr(const double* v1, const double* v2)
 {
-	const float dx = v2[0] - v1[0];
-	const float dz = v2[2] - v1[2];
+	const double dx = v2[0] - v1[0];
+	const double dz = v2[2] - v1[2];
 	return dx*dx + dz*dz;
 }
 
 /// Normalizes the vector.
 ///  @param[in,out]	v	The vector to normalize. [(x, y, z)]
-inline void dtVnormalize(float* v)
+inline void dtVnormalize(double* v)
 {
-	float d = 1.0f / dtSqrt(dtSqr(v[0]) + dtSqr(v[1]) + dtSqr(v[2]));
+	double d = 1.0 / dtSqrt(dtSqr(v[0]) + dtSqr(v[1]) + dtSqr(v[2]));
 	v[0] *= d;
 	v[1] *= d;
 	v[2] *= d;
@@ -278,10 +278,10 @@ inline void dtVnormalize(float* v)
 ///
 /// Basically, this function will return true if the specified points are 
 /// close enough to eachother to be considered colocated.
-inline bool dtVequal(const float* p0, const float* p1)
+inline bool dtVequal(const double* p0, const double* p1)
 {
-	static const float thr = dtSqr(1.0f/16384.0f);
-	const float d = dtVdistSqr(p0, p1);
+	static const double thr = dtSqr(1.0/16384.0);
+	const double d = dtVdistSqr(p0, p1);
 	return d < thr;
 }
 
@@ -291,7 +291,7 @@ inline bool dtVequal(const float* p0, const float* p1)
 /// @return The dot product on the xz-plane.
 ///
 /// The vectors are projected onto the xz-plane, so the y-values are ignored.
-inline float dtVdot2D(const float* u, const float* v)
+inline double dtVdot2D(const double* u, const double* v)
 {
 	return u[0]*v[0] + u[2]*v[2];
 }
@@ -302,7 +302,7 @@ inline float dtVdot2D(const float* u, const float* v)
 /// @return The dot product on the xz-plane.
 ///
 /// The vectors are projected onto the xz-plane, so the y-values are ignored.
-inline float dtVperp2D(const float* u, const float* v)
+inline double dtVperp2D(const double* u, const double* v)
 {
 	return u[2]*v[0] - u[0]*v[2];
 }
@@ -316,12 +316,12 @@ inline float dtVperp2D(const float* u, const float* v)
 ///  @param[in]		b		Vertex B. [(x, y, z)]
 ///  @param[in]		c		Vertex C. [(x, y, z)]
 /// @return The signed xz-plane area of the triangle.
-inline float dtTriArea2D(const float* a, const float* b, const float* c)
+inline double dtTriArea2D(const double* a, const double* b, const double* c)
 {
-	const float abx = b[0] - a[0];
-	const float abz = b[2] - a[2];
-	const float acx = c[0] - a[0];
-	const float acz = c[2] - a[2];
+	const double abx = b[0] - a[0];
+	const double abz = b[2] - a[2];
+	const double acx = c[0] - a[0];
+	const double acz = c[2] - a[2];
 	return acx*abz - abx*acz;
 }
 
@@ -349,8 +349,8 @@ inline bool dtOverlapQuantBounds(const unsigned short amin[3], const unsigned sh
 ///  @param[in]		bmax	Maximum bounds of box B. [(x, y, z)]
 /// @return True if the two AABB's overlap.
 /// @see dtOverlapQuantBounds
-inline bool dtOverlapBounds(const float* amin, const float* amax,
-							const float* bmin, const float* bmax)
+inline bool dtOverlapBounds(const double* amin, const double* amax,
+							const double* bmin, const double* bmax)
 {
 	bool overlap = true;
 	overlap = (amin[0] > bmax[0] || amax[0] < bmin[0]) ? false : overlap;
@@ -365,8 +365,8 @@ inline bool dtOverlapBounds(const float* amin, const float* amax,
 ///  @param[in]		a		Vertex A of triangle ABC. [(x, y, z)]
 ///  @param[in]		b		Vertex B of triangle ABC. [(x, y, z)]
 ///  @param[in]		c		Vertex C of triangle ABC. [(x, y, z)]
-void dtClosestPtPointTriangle(float* closest, const float* p,
-							  const float* a, const float* b, const float* c);
+void dtClosestPtPointTriangle(double* closest, const double* p,
+							  const double* a, const double* b, const double* c);
 
 /// Derives the y-axis height of the closest point on the triangle from the specified reference point.
 ///  @param[in]		p		The reference point from which to test. [(x, y, z)]
@@ -374,35 +374,35 @@ void dtClosestPtPointTriangle(float* closest, const float* p,
 ///  @param[in]		b		Vertex B of triangle ABC. [(x, y, z)]
 ///  @param[in]		c		Vertex C of triangle ABC. [(x, y, z)]
 ///  @param[out]	h		The resulting height.
-bool dtClosestHeightPointTriangle(const float* p, const float* a, const float* b, const float* c, float& h);
+bool dtClosestHeightPointTriangle(const double* p, const double* a, const double* b, const double* c, double& h);
 
-bool dtIntersectSegmentPoly2D(const float* p0, const float* p1,
-							  const float* verts, int nverts,
-							  float& tmin, float& tmax,
+bool dtIntersectSegmentPoly2D(const double* p0, const double* p1,
+							  const double* verts, int nverts,
+							  double& tmin, double& tmax,
 							  int& segMin, int& segMax);
 
-bool dtIntersectSegSeg2D(const float* ap, const float* aq,
-						 const float* bp, const float* bq,
-						 float& s, float& t);
+bool dtIntersectSegSeg2D(const double* ap, const double* aq,
+						 const double* bp, const double* bq,
+						 double& s, double& t);
 
 /// Determines if the specified point is inside the convex polygon on the xz-plane.
 ///  @param[in]		pt		The point to check. [(x, y, z)]
 ///  @param[in]		verts	The polygon vertices. [(x, y, z) * @p nverts]
 ///  @param[in]		nverts	The number of vertices. [Limit: >= 3]
 /// @return True if the point is inside the polygon.
-bool dtPointInPolygon(const float* pt, const float* verts, const int nverts);
+bool dtPointInPolygon(const double* pt, const double* verts, const int nverts);
 
-bool dtDistancePtPolyEdgesSqr(const float* pt, const float* verts, const int nverts,
-							float* ed, float* et);
+bool dtDistancePtPolyEdgesSqr(const double* pt, const double* verts, const int nverts,
+							double* ed, double* et);
 
-float dtDistancePtSegSqr2D(const float* pt, const float* p, const float* q, float& t);
+double dtDistancePtSegSqr2D(const double* pt, const double* p, const double* q, double& t);
 
 /// Derives the centroid of a convex polygon.
 ///  @param[out]	tc		The centroid of the polgyon. [(x, y, z)]
 ///  @param[in]		idx		The polygon indices. [(vertIndex) * @p nidx]
 ///  @param[in]		nidx	The number of indices in the polygon. [Limit: >= 3]
 ///  @param[in]		verts	The polygon vertices. [(x, y, z) * vertCount]
-void dtCalcPolyCenter(float* tc, const unsigned short* idx, int nidx, const float* verts);
+void dtCalcPolyCenter(double* tc, const unsigned short* idx, int nidx, const double* verts);
 
 /// Determines if the two convex polygons overlap on the xz-plane.
 ///  @param[in]		polya		Polygon A vertices.	[(x, y, z) * @p npolya]
@@ -410,8 +410,8 @@ void dtCalcPolyCenter(float* tc, const unsigned short* idx, int nidx, const floa
 ///  @param[in]		polyb		Polygon B vertices.	[(x, y, z) * @p npolyb]
 ///  @param[in]		npolyb		The number of vertices in polygon B.
 /// @return True if the two polygons overlap.
-bool dtOverlapPolyPoly2D(const float* polya, const int npolya,
-						 const float* polyb, const int npolyb);
+bool dtOverlapPolyPoly2D(const double* polya, const int npolya,
+						 const double* polyb, const int npolyb);
 
 /// @}
 /// @name Miscellanious functions.
@@ -476,14 +476,14 @@ inline void dtSwapEndian(int* v)
 	dtSwapByte(x+0, x+3); dtSwapByte(x+1, x+2);
 }
 
-inline void dtSwapEndian(float* v)
+inline void dtSwapEndian(double* v)
 {
 	unsigned char* x = (unsigned char*)v;
 	dtSwapByte(x+0, x+3); dtSwapByte(x+1, x+2);
 }
 
-void dtRandomPointInConvexPoly(const float* pts, const int npts, float* areas,
-							   const float s, const float t, float* out);
+void dtRandomPointInConvexPoly(const double* pts, const int npts, double* areas,
+							   const double s, const double t, double* out);
 
 /// @}
 
@@ -496,7 +496,7 @@ void dtRandomPointInConvexPoly(const float* pts, const int npts, float* areas,
 
 /**
 
-@fn float dtTriArea2D(const float* a, const float* b, const float* c)
+@fn double dtTriArea2D(const double* a, const double* b, const double* c)
 @par
 
 The vertices are projected onto the xz-plane, so the y-values are ignored.
